initial cleanup including prof resolver split
